+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
-    "files.eol": "\n",
-    "editor.detectIndentation": false,
-    "editor.formatOnSave": false,
-    "editor.tabSize": 2
+  "files.eol": "\n",
+  "editor.detectIndentation": false,
+  "editor.formatOnSave": false,
+  "editor.tabSize": 2
 }
