+++ src/views/MovieView.tsx
@@ -10,7 +10,6 @@ export function MovieView() {
   const store = useWatchedContext();
 
   useEffect(() => {
-    throw new Error("Hi");
     (async () => {
       setStreamUrl(mediaPortable && (await getStream(mediaPortable)));
     })();
